- updating from the previous auction house version should now be working - you can specify a decimal point now when you put a suffix like k or m: "/ah sell 1.2m"

+++ src/main/java/me/elaineqheart/auctionHouse/data/persistentStorage/ItemNoteStorage.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 
 public class ItemNoteStorage {
 
-    public static void createNote(Player p, ItemStack item, int price, boolean isBINAuction) {
+    public static void createNote(Player p, ItemStack item, double price, boolean isBINAuction) {
         //if(r()) RedisNoteStorage.createNote(p, item, price); else
         JsonNoteStorage.createNote(p, item, price, isBINAuction);
     }
